requirements.txt
alembic==1.17.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Dance==7.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
URLObject==3.0.0
Werkzeug==3.1.3
